refactor and put the config in its own component
